requirements.txt
customtkinter
speechrecognition
pyttsx3
pyaudio
psutil
pywhatkit
google-generativeai
pillow
screen-brightness-control